+++ examples/node-gtk/gtk-3-editor/main.ts
@@ -1,3 +1,4 @@
+
 import gi from './@types/node-gtk'
 import * as Gtk from './@types/node-gtk-3.0'
 import * as GtkSource from './@types/node-gtksource-3.0'